Introduce the function strip_path_suffix()
The function strip_path_suffix() will try to split the given path into
prefix/suffix.  The suffix has to be passed to the function, and if the
path matches, the prefix is set.

Arbitrary runs of directory separators ("slashes") are assumed identical.

Example:

	strip_path_suffix("C:\\msysgit/\\libexec\\git-core",
		"libexec///git-core", &prefix)

will set prefix to "C:\\msysgit" and return 0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>