Teach rebase about preserving merges while in interactive mode
It has been quite a while since I tried my hand at this... But
with Git for Windows' fork needing to hold several topic branches,
I felt the need for something practical. Something like 'git rebase
--interactive -preserve-merges' that can handle fixup commits.

The major obstacle was the need to implement our own topological
order.  Why? Because rev-list's topo-order is overly eager to
continue with the *last* parent after inspecting a merge commit,
whereas we want to continue with the *first* parent so that we have
longer runs of "pick" and "merge" commands without having to jump
around (if you're a Badger, you may sing and dance at this point).

This is how we do it: we first make a list of tip commits (i.e.
commits at the end of afore-mentioned pick & merge runs), in the
order in which we want to build the edit script.

For each tip, we then build the edit script in *reverse* order,
traversing from each commit to its first parent. Whenever we
encounter a parent commit that is already in the edit script, we
remember to label it and stop. Upon encountering a first parent
commit that is not to be rebased at all, we are done with that
particular tip, too.

We had to add three new commands for this to work:

	label <name>: make a temporary ref for the now-current commit
	goto <name>: jump to a previously labeled temporary ref
	merge -c <commit> <parents>: perform a merge

Technically, we could have done all of these as "exec" lines, but
they would have looked horribly ugly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>